commitlint: separate tests in 2 files
Strictly speaking, there's one test that is checking that
one of the commitlint's native rules works (not a plugin
of ours).
